test: the address API-related suite re-structuring
