+++ packages/integration-tests/tests/api/common/endpoints/tokens-endpoints.test.ts
@@ -1,11 +1,11 @@
 import * as request from "supertest";
 import { setTimeout } from "timers/promises";
 
-import { environment } from "../../../src/config";
-import { localConfig } from "../../../src/config";
-import { Buffer, Token, TransactionsType, Wallets } from "../../../src/entities";
-import { Helper } from "../../../src/helper";
-import { Playbook } from "../../../src/playbook/playbook";
+import { environment } from "../../../../src/config";
+import { localConfig } from "../../../../src/config";
+import { Buffer, Token, TransactionsType, Wallets } from "../../../../src/entities";
+import { Helper } from "../../../../src/helper";
+import { Playbook } from "../../../../src/playbook/playbook";
 
 describe("Tokens", () => {
   jest.setTimeout(localConfig.standardTimeout);
