+++ packages/integration-tests/tests/api/common/nft.test.ts
@@ -26,29 +26,5 @@ describe("NFTs", () => {
       deployedToken = await playbook.deployNFTtoL2();
       expect(deployedToken).toContain(Logger.txHashStartsWith);
     });
-
-    //@id1457
-    it("Verify deployed to L2 NFT via /address/{address}", async () => {
-      await setTimeout(localConfig.extendedPause);
-      const apiRoute = `/address/${deployedToken}`;
-
-      return request(environment.blockExplorerAPI)
-        .get(apiRoute)
-        .expect(200)
-        .expect((res) =>
-          expect(res.body).toEqual(
-            expect.objectContaining({
-              type: "contract",
-              address: deployedToken,
-              balances: {
-                [`${deployedToken}`]: {
-                  balance: "1",
-                  token: null,
-                },
-              },
-            })
-          )
-        );
-    });
   });
 });
