+++ packages/integration-tests/tests/api/transactions/transaction.test.ts
@@ -135,19 +135,6 @@ describe("Transactions", () => {
       expect(contract).toContain(Logger.txHashStartsWith);
     });
 
-    //@id1449
-    it("Verify the deployed Greeter contract via /address/{address}", async () => {
-      const apiRoute = `/address/${contract}`;
-
-      await setTimeout(localConfig.standardPause); //works unstable without timeout
-
-      return request(environment.blockExplorerAPI)
-        .get(apiRoute)
-        .expect(200)
-        .expect((res) => expect(res.body).toStrictEqual(expect.objectContaining({ address: contract })))
-        .expect((res) => expect(res.body).toStrictEqual(expect.objectContaining({ balances: {} })));
-    });
-
     //@id604
     it("Use the Greeter contract - execute SetGreeting", async () => {
       const executedContract = await playbook.useGreeter();
@@ -210,110 +197,6 @@ describe("Transactions", () => {
     });
   });
 
-  describe("Transactions for the /address/{address} endpoints", () => {
-    jest.setTimeout(localConfig.extendedTimeout);
-
-    //@id1510
-    it("Verify the transaction via /address/{address}/logs", async () => {
-      contract = await helper.getStringFromFile(bufferRoute + Buffer.greeterL2);
-      txHash = await helper.getStringFromFile(bufferRoute + Buffer.executeGreeterTx);
-
-      const apiRoute = `/address/${contract}/logs`;
-      const decapitalizedAddress = apiRoute.slice(1).toLowerCase();
-
-      await setTimeout(localConfig.standardPause); //works unstable without timeout
-
-      return request(environment.blockExplorerAPI)
-        .get(apiRoute)
-        .expect(200)
-        .expect((res) => expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ address: contract })))
-        .expect((res) => expect(Array.isArray(res.body.items[0].topics)).toStrictEqual(true))
-        .expect((res) => expect(typeof res.body.items[0].data).toStrictEqual("string"))
-        .expect((res) => expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ transactionHash: txHash })))
-        .expect((res) => expect(typeof res.body.items[0].transactionIndex).toStrictEqual("number"))
-        .expect((res) => expect(typeof res.body.items[0].logIndex).toStrictEqual("number"))
-        .expect((res) => expect(typeof res.body.items[0].timestamp).toStrictEqual("string"))
-        .expect((res) => expect(res.body.meta).toStrictEqual(expect.objectContaining({ totalItems: 1 })))
-        .expect((res) => expect(res.body.meta).toStrictEqual(expect.objectContaining({ itemCount: 1 })))
-        .expect((res) => expect(res.body.meta).toStrictEqual(expect.objectContaining({ itemsPerPage: 10 })))
-        .expect((res) => expect(res.body.meta).toStrictEqual(expect.objectContaining({ totalPages: 1 })))
-        .expect((res) => expect(res.body.meta).toStrictEqual(expect.objectContaining({ currentPage: 1 })))
-        .expect((res) =>
-          expect(res.body.links).toStrictEqual(expect.objectContaining({ first: `${decapitalizedAddress}?limit=10` }))
-        )
-        .expect((res) => expect(res.body.links).toStrictEqual(expect.objectContaining({ previous: "" })))
-        .expect((res) => expect(res.body.links).toStrictEqual(expect.objectContaining({ next: "" })))
-        .expect((res) =>
-          expect(res.body.links).toStrictEqual(
-            expect.objectContaining({ last: `${decapitalizedAddress}?page=1&limit=10` })
-          )
-        );
-    });
-
-    //@id1509
-    it("Verify the transaction via /address/{address}/transfers", async () => {
-      contract = await helper.getStringFromFile(bufferRoute + Buffer.paymaster);
-      emptyWallet = await helper.getStringFromFile(bufferRoute + Buffer.emptyWalletAddress);
-      txHash = await helper.getStringFromFile(bufferRoute + Buffer.paymasterTx);
-
-      const customTokenAddress = await helper.getStringFromFile(bufferRoute + Buffer.customToken);
-      const apiRoute = `/address/${contract}/transfers`;
-      const decapitalizedAddress = apiRoute.slice(1).toLowerCase();
-
-      await setTimeout(localConfig.standardPause); //works unstable without timeout
-
-      return request(environment.blockExplorerAPI)
-        .get(apiRoute)
-        .expect(200)
-        .expect((res) => expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ from: emptyWallet })))
-        .expect((res) => expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ to: contract })))
-        .expect((res) => expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ transactionHash: txHash })))
-        .expect((res) => expect(typeof res.body.items[0].timestamp).toStrictEqual("string"))
-        .expect((res) => expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ amount: "1" })))
-        .expect((res) =>
-          expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ tokenAddress: customTokenAddress }))
-        )
-        .expect((res) => expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ type: "transfer" })))
-        .expect((res) => expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ fields: null })))
-        .expect((res) => expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ fields: null })))
-        .expect((res) =>
-          expect(res.body.items[0].token).toStrictEqual(expect.objectContaining({ l2Address: customTokenAddress }))
-        )
-        .expect((res) => expect(res.body.items[0].token).toStrictEqual(expect.objectContaining({ l1Address: null })))
-        .expect((res) => expect(res.body.items[0].token).toStrictEqual(expect.objectContaining({ symbol: "MyToken" })))
-        .expect((res) => expect(res.body.items[0].token).toStrictEqual(expect.objectContaining({ name: "MyToken" })))
-        .expect((res) => expect(res.body.items[0].token).toStrictEqual(expect.objectContaining({ decimals: 18 })))
-        .expect((res) =>
-          expect(res.body.items[1]).toStrictEqual(expect.objectContaining({ from: Wallets.richWalletAddress }))
-        )
-        .expect((res) => expect(res.body.items[1]).toStrictEqual(expect.objectContaining({ to: contract })))
-        .expect((res) => expect(typeof res.body.items[1].timestamp).toStrictEqual("string"))
-        .expect((res) =>
-          expect(res.body.items[1]).toStrictEqual(expect.objectContaining({ amount: "30000000000000000" }))
-        )
-        .expect((res) =>
-          expect(res.body.items[1]).toStrictEqual(expect.objectContaining({ tokenAddress: Token.ETHER_ERC20_Address }))
-        )
-        .expect((res) => expect(res.body.items[1]).toStrictEqual(expect.objectContaining({ type: "transfer" })))
-        .expect((res) => expect(res.body.items[1]).toStrictEqual(expect.objectContaining({ fields: null })))
-        .expect((res) => expect(res.body.items[1]).toStrictEqual(expect.objectContaining({ fields: null })))
-        .expect((res) => expect(res.body.meta).toStrictEqual(expect.objectContaining({ totalItems: 2 })))
-        .expect((res) => expect(res.body.meta).toStrictEqual(expect.objectContaining({ itemCount: 2 })))
-        .expect((res) => expect(res.body.meta).toStrictEqual(expect.objectContaining({ itemsPerPage: 10 })))
-        .expect((res) => expect(res.body.meta).toStrictEqual(expect.objectContaining({ totalPages: 1 })))
-        .expect((res) => expect(res.body.meta).toStrictEqual(expect.objectContaining({ currentPage: 1 })))
-        .expect((res) =>
-          expect(res.body.links).toStrictEqual(expect.objectContaining({ first: `${decapitalizedAddress}?limit=10` }))
-        )
-        .expect((res) => expect(res.body.links).toStrictEqual(expect.objectContaining({ previous: "" })))
-        .expect((res) => expect(res.body.links).toStrictEqual(expect.objectContaining({ next: "" })))
-        .expect((res) =>
-          expect(res.body.links).toStrictEqual(
-            expect.objectContaining({ last: `${decapitalizedAddress}?page=1&limit=10` })
-          )
-        );
-    });
-  });
   describe("Transactions for the /transactions endpoint", () => {
     jest.setTimeout(localConfig.extendedTimeout);
